+++ assets/scenes/menu/menu.fire
@@ -171,7 +171,7 @@
       "array": [
         0,
         0,
-        234.6928844255829,
+        231.22878281044513,
         0,
         0,
         0,
