+++ assets/scripts/controller/SelectArenaKeyboardController.ts
@@ -1,6 +1,5 @@
 import KeyboardController from "./KeyboardController";
-import ArenaOptions from "../ArenaOptions"
-
+import ArenaOptions from "../arena_select/ArenaOptions";
 const {ccclass, property} = cc._decorator;
 
 @ccclass
